fix: chart, form save
